[IMP] sale: changes after tde review
