+++ addons/sale/__openerp__.py
@@ -43,28 +43,28 @@ The Dashboard for the Sales Manager will include
     'website': 'https://www.odoo.com/page/crm',
     'depends': ['sales_team', 'account', 'procurement', 'report', 'web_tour'],
     'data': [
-        'wizard/sale_make_invoice_advance_views.xml',
-        'security/sale_security.xml',
         'security/ir.model.access.csv',
-        'data/sequence_data.xml',
+        'security/sale_security.xml',
+        'wizard/sale_make_invoice_advance_views.xml',
         'report/sale_report.xml',
-        'data/sale_data.xml',
-        'views/sale_views.xml',
-        'views/account_invoice_views.xml',
         'report/sale_report_views.xml',
         'report/account_invoice_report_views.xml',
+        'report/sale_report_templates.xml',
+        'report/invoice_report_templates.xml',
+        'views/sale_views.xml',
+        'views/account_invoice_views.xml',
         'views/sales_team_views.xml',
         'views/res_partner_views.xml',
-        'data/tour_data.xml',
-        'data/mail_template_data.xml',
         'views/res_config_views.xml',
-        'report/sale_report_templates.xml',
         'views/sale_templates.xml',
-        'report/invoice_report_templates.xml',
         'views/sale_layout_category_views.xml',
         'views/product_product_views.xml',
         'views/product_template_views.xml',
         'views/res_company_views.xml',
+        'data/mail_template_data.xml',
+        'data/sale_data.xml',
+        'data/sequence_data.xml',
+        'data/tour_data.xml',
     ],
     'demo': ['data/sale_demo.xml',
              'data/product_demo.xml',
