+++ addons/sale/models/__init__.py
@@ -1,14 +1,15 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+import account_config_settings
 import account_invoice
 import procurement
 import product_product
 import product_template
 import res_company
-import res_config
 import res_partner
 import sale
 import sale_analytic
+import sale_config_settings
 import sale_layout
 import sales_team
